requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
slowapi
pydantic
pydub
python-multipart
requests
